Add CNN to prediction range
Put CNN in a function so both encoding and decoding ranges can use. 
Add CNN to prediction range. After predicting the value for current time step, input the last 24 values before current time step to CNN and use the output of CNN as the input for LSTM for next time_step's prediction.